Made type of getIterations() equal to the type of the variable

+++ scenario_ref/include/optimization/RandomOptimizer.h
@@ -23,7 +23,7 @@ public:
 
     nlohmann::json optimize() override;
 
-    int getIterations() {
+    size_t getIterations() {
         return iterations;
     }
